ai-models/requirements.txt
flask==3.0.0
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.1.0
scikit-learn==1.3.2
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
